add presentation jupyter notebook
